% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{activity}
\alias{activity}
\title{activity}
\format{
A data frame with 825 rows and 45 variables:
\describe{
  \item{\code{iati_identifier}}{character COLUMN_DESCRIPTION}
  \item{\code{reporting_org}}{character COLUMN_DESCRIPTION}
  \item{\code{reporting_org_ref}}{character COLUMN_DESCRIPTION}
  \item{\code{reporting_org_type}}{character COLUMN_DESCRIPTION}
  \item{\code{reporting_org_secondary_reporter}}{logical COLUMN_DESCRIPTION}
  \item{\code{title_eng}}{character COLUMN_DESCRIPTION}
  \item{\code{title_fr}}{character COLUMN_DESCRIPTION}
  \item{\code{title_es}}{character COLUMN_DESCRIPTION}
  \item{\code{title_zh}}{character COLUMN_DESCRIPTION}
  \item{\code{title_ru}}{character COLUMN_DESCRIPTION}
  \item{\code{title_ar}}{character COLUMN_DESCRIPTION}
  \item{\code{description_type_1}}{character COLUMN_DESCRIPTION}
  \item{\code{description_eng_1}}{character COLUMN_DESCRIPTION}
  \item{\code{descriptionv_fr_1}}{character COLUMN_DESCRIPTION}
  \item{\code{description_type_2}}{character COLUMN_DESCRIPTION}
  \item{\code{description_eng_2}}{character COLUMN_DESCRIPTION}
  \item{\code{descriptionv_fr_2}}{character COLUMN_DESCRIPTION}
  \item{\code{activity_status_code}}{character COLUMN_DESCRIPTION}
  \item{\code{activity_date_1}}{character COLUMN_DESCRIPTION}
  \item{\code{activity_date_type_1}}{character COLUMN_DESCRIPTION}
  \item{\code{activity_date_2}}{character COLUMN_DESCRIPTION}
  \item{\code{activity_date_type_2}}{character COLUMN_DESCRIPTION}
  \item{\code{activity_date_3}}{character COLUMN_DESCRIPTION}
  \item{\code{activity_date_type_3}}{character COLUMN_DESCRIPTION}
  \item{\code{activity_date_4}}{character COLUMN_DESCRIPTION}
  \item{\code{activity_date_type_4}}{character COLUMN_DESCRIPTION}
  \item{\code{contact_info_type}}{character COLUMN_DESCRIPTION}
  \item{\code{contact_info_org}}{character COLUMN_DESCRIPTION}
  \item{\code{contact_info_email}}{character COLUMN_DESCRIPTION}
  \item{\code{contact_info_website}}{character COLUMN_DESCRIPTION}
  \item{\code{contact_info_mailing_address}}{character COLUMN_DESCRIPTION}
  \item{\code{activity_scope_code}}{character COLUMN_DESCRIPTION}
  \item{\code{recipient_country_code}}{character COLUMN_DESCRIPTION}
  \item{\code{recipient_country_pct}}{logical COLUMN_DESCRIPTION}
  \item{\code{recipient_country}}{logical COLUMN_DESCRIPTION}
  \item{\code{recipient_region_code}}{character COLUMN_DESCRIPTION}
  \item{\code{recipient_region_vocabulary}}{character COLUMN_DESCRIPTION}
  \item{\code{recipient_region_vocabulary_url}}{character COLUMN_DESCRIPTION}
  \item{\code{recipient_region_pct}}{logical COLUMN_DESCRIPTION}
  \item{\code{recipient_region}}{character COLUMN_DESCRIPTION}
  \item{\code{collaboration_type_code}}{character COLUMN_DESCRIPTION}
  \item{\code{default_flow_type_code}}{character COLUMN_DESCRIPTION}
  \item{\code{default_finance_type_code}}{character COLUMN_DESCRIPTION}
  \item{\code{default_tied_status_code}}{character COLUMN_DESCRIPTION}
  \item{\code{capital_spend}}{character COLUMN_DESCRIPTION} 
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
activity
}
\description{
Records of all activities since UNHCR Programme 2016 in IATI format
}
\keyword{datasets}
