% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{budget}
\alias{budget}
\title{Budget}
\format{
A data frame with 840 rows and 8 variables:
\describe{
  \item{\code{iati_identifier}}{character COLUMN_DESCRIPTION}
  \item{\code{budget_type}}{character COLUMN_DESCRIPTION}
  \item{\code{budget_status}}{character COLUMN_DESCRIPTION}
  \item{\code{budget_period_start}}{character COLUMN_DESCRIPTION}
  \item{\code{budget_period_end}}{character COLUMN_DESCRIPTION}
  \item{\code{budget_currency}}{character COLUMN_DESCRIPTION}
  \item{\code{budget_value_date}}{character COLUMN_DESCRIPTION}
  \item{\code{budget_value}}{character COLUMN_DESCRIPTION} 
}
}
\source{
\url{https://reporting.unhcr.org/iati}
}
\usage{
budget
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
