% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codeActivityDateType}
\alias{codeActivityDateType}
\title{codeActivityDateType}
\format{
A data frame with 4 rows and 6 variables:
\describe{
  \item{\code{code}}{double code}
  \item{\code{name}}{character name}
  \item{\code{description}}{character description}
  \item{\code{category}}{double  category}
  \item{\code{url}}{character url}
  \item{\code{status}}{character status}
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/codelists/}
}
\usage{
codeActivityDateType
}
\description{
Types of date for activities.
There are many different business models and dates that can be used to describe
 the start and end of activities. It is recommended that each publisher
 adopts their own consistent approach that provides users with a meaningful
 indication of the lifespan of an activity.
}
\examples{
{
knitr::kable(head(codeActivityDateType, 10))
}
}
\keyword{datasets}
