% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codeActivityStatus}
\alias{codeActivityStatus}
\title{codeActivityStatus}
\format{
A data frame with 6 rows and 6 variables:
\describe{
  \item{\code{code}}{double code}
  \item{\code{name}}{character name}
  \item{\code{description}}{character description}
  \item{\code{category}}{double  category}
  \item{\code{url}}{character url}
  \item{\code{status}}{character status}
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/codelists/}
}
\usage{
codeActivityStatus
}
\description{
Lifecycle status of the activity from pipeline to completion
}
\examples{
{
knitr::kable(head(codeActivityStatus, 10))
}
}
\keyword{datasets}
