% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codeAidTypeVocabulary}
\alias{codeAidTypeVocabulary}
\title{codeAidTypeVocabulary}
\format{
A data frame with 4 rows and 6 variables:
\describe{
  \item{\code{code}}{double code}
  \item{\code{name}}{character name}
  \item{\code{description}}{character description}
  \item{\code{category}}{double  category}
  \item{\code{url}}{character url}
  \item{\code{status}}{character status}
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/codelists/}
}
\usage{
codeAidTypeVocabulary
}
\description{
The AidTypeVocabulary codelist defines a range of external
codelists which themselves provide codes and descriptions for aid type.
This is a Non-Core codelist.
}
\examples{
{
knitr::kable(head(codeAidTypeVocabulary, 10))
}
}
\keyword{datasets}
