% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codeBudgetIdentifier}
\alias{codeBudgetIdentifier}
\title{codeBudgetIdentifier}
\format{
A data frame with 152 rows and 6 variables:
\describe{
  \item{\code{code}}{double code}
  \item{\code{name}}{character name}
  \item{\code{description}}{character description}
  \item{\code{category}}{double  category}
  \item{\code{url}}{character url}
  \item{\code{status}}{character status}
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/codelists/}
}
\usage{
codeBudgetIdentifier
}
\description{
IATI Functional and Administrative Common Code :
One of several possible Budget Identifier Vocabularies.
As of version 2.03 this codelist has been deprecated.
  This is a Non-Core codelist.
}
\examples{
{
knitr::kable(head(codeBudgetIdentifier, 10))
}
}
\keyword{datasets}
