% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codeBudgetIdentifierSector}
\alias{codeBudgetIdentifierSector}
\title{codeBudgetIdentifierSector}
\format{
A data frame with 31 rows and 6 variables:
\describe{
  \item{\code{code}}{double code}
  \item{\code{name}}{character name}
  \item{\code{description}}{character description}
  \item{\code{category}}{double  category}
  \item{\code{url}}{character url}
  \item{\code{status}}{character status}
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/codelists/}
}
\usage{
codeBudgetIdentifierSector
}
\description{
This codelists exists to group the Budget Identifier codelist
into sectors. It is not used as a codelist in its own right.
  This is a Non-Core codelist.
}
\examples{
{
knitr::kable(head(codeBudgetIdentifierSector, 10))
}
}
\keyword{datasets}
