% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codeCashandVoucherModalities}
\alias{codeCashandVoucherModalities}
\title{codeCashandVoucherModalities}
\format{
A data frame with 2 rows and 6 variables:
\describe{
  \item{\code{code}}{double code}
  \item{\code{name}}{character name}
  \item{\code{description}}{character description}
  \item{\code{category}}{double  category}
  \item{\code{url}}{character url}
  \item{\code{status}}{character status}
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/codelists/}
}
\usage{
codeCashandVoucherModalities
}
\description{
This codelist has been created by IATI following agreements
and recommendations of the Tracking Cash and Voucher Assistance (CVA)
Working Group. Definitions of the codes have been aligned with the
CaLP Glossary:http://www.cashlearning.org/resources/glossary.
  This is a Non-Core codelist.
}
\examples{
{
knitr::kable(head(codeCashandVoucherModalities, 10))
}
}
\keyword{datasets}
