% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codeCollaborationType}
\alias{codeCollaborationType}
\title{codeCollaborationType}
\format{
A data frame with 7 rows and 6 variables:
\describe{
  \item{\code{code}}{double code}
  \item{\code{name}}{character name}
  \item{\code{description}}{character description}
  \item{\code{category}}{double  category}
  \item{\code{url}}{character url}
  \item{\code{status}}{character status}
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/codelists/collaborationtype/}
}
\usage{
codeCollaborationType
}
\description{
OECD DAC classification used to determine the character of
    resource flows (bilateral or multilateral).
  External URL: http://www.oecd.org/dac/stats/dacandcrscodelists.htm
}
\examples{
{
knitr::kable(head(codeCollaborationType, 10))
}
}
\keyword{datasets}
