% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codeCountry}
\alias{codeCountry}
\title{codeCountry}
\format{
A data frame with 251 rows and 6 variables:
\describe{
  \item{\code{code}}{double code}
  \item{\code{name}}{character name}
  \item{\code{description}}{character description}
  \item{\code{category}}{double  category}
  \item{\code{url}}{character url}
  \item{\code{status}}{character status}
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/codelists/}
}
\usage{
codeCountry
}
\description{
The Country codelist is generated from the ISO 3166-1 part
 of the ISO 3166 standard. The standard makes allowance, alongside
  the officially assigned codes, for code elements to be expanded by using
  either reserved codes or user-assigned codes.
  IATI currently defines additional codes in the XA -XZ range.
    External URL: http://www.iso.org/iso/home/standards/country_codes.htm
}
\examples{
{
knitr::kable(head(codeCountry, 10))
}
}
\keyword{datasets}
