% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codeCurrency}
\alias{codeCurrency}
\title{codeCurrency}
\format{
A data frame with 174 rows and 6 variables:
\describe{
  \item{\code{code}}{double code}
  \item{\code{name}}{character name}
  \item{\code{description}}{character description}
  \item{\code{category}}{double  category}
  \item{\code{url}}{character url}
  \item{\code{status}}{character status}
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/codelists/}
}
\usage{
codeCurrency
}
\description{
ISO 4217 Currency used for all transactions and budgets
  External URL: http://www.iso.org/iso/home/standards/currency_codes.htm
}
\examples{
{
knitr::kable(head(codeCurrency, 10))
}
}
\keyword{datasets}
