% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codeDescriptionType}
\alias{codeDescriptionType}
\title{codeDescriptionType}
\format{
A data frame with 4 rows and 6 variables:
\describe{
  \item{\code{code}}{double code}
  \item{\code{name}}{character name}
  \item{\code{description}}{character description}
  \item{\code{category}}{double  category}
  \item{\code{url}}{character url}
  \item{\code{status}}{character status}
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/codelists/}
}
\usage{
codeDescriptionType
}
\description{
Activity decription types. (General, objectives, etc)
}
\examples{
{
knitr::kable(head(codeDescriptionType, 10))
}
}
\keyword{datasets}
