% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codeEarmarkingCategory}
\alias{codeEarmarkingCategory}
\title{codeEarmarkingCategory}
\format{
A data frame with 4 rows and 6 variables:
\describe{
  \item{\code{code}}{double code}
  \item{\code{name}}{character name}
  \item{\code{description}}{character description}
  \item{\code{category}}{double  category}
  \item{\code{url}}{character url}
  \item{\code{status}}{character status}
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/codelists/earmarkingcategory/}
}
\usage{
codeEarmarkingCategory
}
\description{
This codelist has been created by IATI and is derived from the
Grand Bargain Earmarking Modality codelist.
See: https://reliefweb.int/sites/reliefweb.int/files/resources/Grand_Bargain_final_22_May_FINAL-2.pdf
}
\examples{
{
knitr::kable(head(codeEarmarkingCategory, 10))
}
}
\keyword{datasets}
