% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codeFlowType}
\alias{codeFlowType}
\title{codeFlowType}
\format{
A data frame with 10 rows and 6 variables:
\describe{
  \item{\code{code}}{double code}
  \item{\code{name}}{character name}
  \item{\code{description}}{character description}
  \item{\code{category}}{double  category}
  \item{\code{url}}{character url}
  \item{\code{status}}{character status}
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/codelists/}
}
\usage{
codeFlowType
}
\description{
DAC/CRS distinction between ODA (official development assistance)
 and other types of resource flow.
   External URL: http://www.oecd.org/dac/stats/dacandcrscodelists.htm
}
\examples{
{
knitr::kable(head(codeFlowType, 10))
}
}
\keyword{datasets}
