% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codeIndicatorMeasure}
\alias{codeIndicatorMeasure}
\title{codeIndicatorMeasure}
\format{
A data frame with 5 rows and 6 variables:
\describe{
  \item{\code{code}}{double code}
  \item{\code{name}}{character name}
  \item{\code{description}}{character description}
  \item{\code{category}}{double  category}
  \item{\code{url}}{character url}
  \item{\code{status}}{character status}
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/codelists/}
}
\usage{
codeIndicatorMeasure
}
\description{
To specify how an indicator is being measured.
This includes qualitative and quantitative values.
}
\examples{
{
knitr::kable(head(codeIndicatorMeasure, 10))
}
}
\keyword{datasets}
