% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codeOrganisationIdentifier}
\alias{codeOrganisationIdentifier}
\title{codeOrganisationIdentifier}
\format{
A data frame with 419 rows and 6 variables:
\describe{
  \item{\code{code}}{double code}
  \item{\code{name}}{character name}
  \item{\code{description}}{character description}
  \item{\code{category}}{double  category}
  \item{\code{url}}{character url}
  \item{\code{status}}{character status}
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/codelists/}
}
\usage{
codeOrganisationIdentifier
}
\description{
As of 1.04 this list is no longer maintained.
 http://support.iatistandard.org/entries/28497976-Retire-the-Organisation-Identifier-codelist#view-post-25368673
   For general guidance about constructing Organisation Identifiers,
   please see http://iatistandard.org/organisation-identifiers/
}
\examples{
{
knitr::kable(head(codeOrganisationIdentifier, 10))
}
}
\keyword{datasets}
