% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codeOrganisationRole}
\alias{codeOrganisationRole}
\title{codeOrganisationRole}
\format{
A data frame with 4 rows and 6 variables:
\describe{
  \item{\code{code}}{double code}
  \item{\code{name}}{character name}
  \item{\code{description}}{character description}
  \item{\code{category}}{double  category}
  \item{\code{url}}{character url}
  \item{\code{status}}{character status}
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/codelists/organisationrole/}
}
\usage{
codeOrganisationRole
}
\description{
IATI codes for the role of an organisation within an activity.
 An organisation can play more than one role within an activity.
   This is a Core codelist.
}
\examples{
{
knitr::kable(head(codeOrganisationRole, 10))
}
}
\keyword{datasets}
