% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codeOrganisationType}
\alias{codeOrganisationType}
\title{codeOrganisationType}
\format{
A data frame with 16 rows and 6 variables:
\describe{
  \item{\code{code}}{double code}
  \item{\code{name}}{character name}
  \item{\code{description}}{character description}
  \item{\code{category}}{double  category}
  \item{\code{url}}{character url}
  \item{\code{status}}{character status}
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/codelists/organisationtype/}
}
\usage{
codeOrganisationType
}
\description{
This is a Non-Core codelist
}
\examples{
{
knitr::kable(head(codeOrganisationType, 10))
}
}
\keyword{datasets}
