% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codePolicyMarker}
\alias{codePolicyMarker}
\title{codePolicyMarker}
\format{
A data frame with 12 rows and 6 variables:
\describe{
  \item{\code{code}}{double code}
  \item{\code{name}}{character name}
  \item{\code{description}}{character description}
  \item{\code{category}}{double  category}
  \item{\code{url}}{character url}
  \item{\code{status}}{character status}
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/codelists/}
}
\usage{
codePolicyMarker
}
\description{
The Policy Marker codelist is derived from the policy markers
 declared by the WP-STAT. The codes themselves are created by IATI.
   External URL: http://www.oecd.org/dac/stats/dacandcrscodelists.htm
}
\examples{
{
knitr::kable(head(codePolicyMarker, 10))
}
}
\keyword{datasets}
