% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codeSectorVocabulary}
\alias{codeSectorVocabulary}
\title{codeSectorVocabulary}
\format{
A data frame with 13 rows and 6 variables:
\describe{
  \item{\code{code}}{double code}
  \item{\code{name}}{character name}
  \item{\code{description}}{character description}
  \item{\code{category}}{double  category}
  \item{\code{url}}{character url}
  \item{\code{status}}{character status}
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/codelists/sectorvocabulary/}
}
\usage{
codeSectorVocabulary
}
\description{
This is a Non-Core codelist.
}
\examples{
{
knitr::kable(head(codeSectorVocabulary, 10))
}
}
\keyword{datasets}
