% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codeUNSDGGoals}
\alias{codeUNSDGGoals}
\title{codeUNSDGGoals}
\format{
A data frame with 17 rows and 6 variables:
\describe{
  \item{\code{code}}{double code}
  \item{\code{name}}{character name}
  \item{\code{description}}{character description}
  \item{\code{category}}{double  category}
  \item{\code{url}}{character url}
  \item{\code{status}}{character status}
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/codelists/}
}
\usage{
codeUNSDGGoals
}
\description{
A value from the top-level list of UN sustainable development
goals (SDGs) (e.g. ‘1’)
  External URL: https://sustainabledevelopment.un.org/?menu=1300
}
\examples{
{
knitr::kable(head(codeUNSDGGoals, 10))
}
}
\keyword{datasets}
