% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_resource_result.R
\name{compare_resource_result}
\alias{compare_resource_result}
\title{compare_resource_result}
\usage{
compare_resource_result(year = 2022, ctr_name = NULL, indicator, outcome)
}
\arguments{
\item{year}{A numeric value or a vector of numeric value to filter on year. Note that data pre-2022 are using a different set of indicators}

\item{ctr_name}{A character vector corresponding to the name of the country.}

\item{indicator}{one specific indicators - label}

\item{outcome}{any of:
"OA1: Access to Territory, Reg. and Documentation", 
"OA2: Status Determination",
"OA3: Protection Policy and Law", 
"OA4: Sexual and Gender-based Violence",
"OA5: Child Protection",
"OA6: Safety and Access to Justice", 
"OA7: Community Engagement and Women's Empowerment", 
"OA8: Well-Being and Basic Needs", 
"OA9: Sustainable Housing and Settlements", 
"OA10: Healthy Lives" ,
"OA11: Education",  
"OA12: Clean Water, Sanitation and Hygiene",
"OA13: Self Reliance, Economic Inclusion and Livelihoods",
"OA14: Voluntary Repatriation and Sustainable Reintegration",
"OA15: Resettlement and Complementary Pathways", 
"OA16: Local Integration and other Local Solutions"}
}
\value{
a graph
}
\description{
compare_resource_result
}
\examples{
# See a list of indicators per area of work
knitr::kable(iati::mapping_indicator |> 
               janitor::clean_names() |>
               dplyr::filter( results_level =="Outcome") |>
               dplyr::select(area_of_work,  indicator) )

compare_resource_result(year =   2022, 
                    indicator = "13.2. Proportion of PoC who self-report positive changes in their income compared to previous year.",
                    outcome = "OA8: Well-Being and Basic Needs")

compare_resource_result(year =   2022, 
                        ctr_name = "Costa Rica",
                    indicator = "13.2. Proportion of PoC who self-report positive changes in their income compared to previous year.",
                    outcome = "OA8: Well-Being and Basic Needs")
}
