% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataParticipating_org}
\alias{dataParticipating_org}
\title{participating_org}
\format{
A data frame with 13990 rows and 7 variables:
\describe{
  \item{\code{iati_identifier}}{character A globally unique identifier for the activity.}
  \item{\code{participating_org_eng}}{character COLUMN_DESCRIPTION}
  \item{\code{participating_org_ref}}{character Machine-readable identification string for the organisation issuing the report. Must be in the format {RegistrationAgency}-{RegistrationNumber} where {RegistrationAgency} is a valid code in the Organisation Registration Agency code list and {RegistrationNumber} is a valid identifier issued by the {RegistrationAgency}. If this is not present then the narrative MUST contain the name of the organisation.}
  \item{\code{participating_org_type}}{double The type of organisation issuing the report. See IATI codelist codeOrganisationType for values.}
  \item{\code{participating_org_role}}{double An IATI code describing the organisation’s role in the activity (donor, agency, etc.). see codeOrganisationRole}
  \item{\code{participating_org_activity_id}}{logical A valid activity identifier published by the participating organisation which points to the activity that it has published to IATI that describes its role in this activity.}
  \item{\code{participating_org_crs_channel_code}}{logical Under CRS++ Reporting Directives this code identifies the implementing agency. Codes ending in ‘00’ are generic and are similar to the OrganisationType code.} 
  \item{\code{participating_org_type_name}}{character COLUMN_DESCRIPTION}
  \item{\code{participating_org_type_description}}{character COLUMN_DESCRIPTION}
  \item{\code{participating_org_role_name}}{character COLUMN_DESCRIPTION}
  \item{\code{participating_org_role_description}}{character COLUMN_DESCRIPTION} 
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/}
}
\usage{
dataParticipating_org
}
\description{
An organisation involved with the activity. May be a donor, fund, agency, etc.
 Specifying the ref identifier is strongly recommended. May contain the organisation name as narrative. 
If the reporting organisation plays a role in the activity it should be repeated here. One organisation may play more than one role (eg, funding and implementing): in such a case each role should be reported and the name of the organisation repeated.
}
\examples{
{
knitr::kable(head(dataParticipating_org, 10))
}
}
\keyword{datasets}
