% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataRelated_activity}
\alias{dataRelated_activity}
\title{dataRelated_activity}
\format{
A data frame  
\describe{
  \item{\code{iati_identifier}}{character A globally unique identifier for the activity.}
  \item{\code{related_activity_ref}}{character COLUMN_DESCRIPTION}
  \item{\code{related_activity_type}}{character COLUMN_DESCRIPTION} 
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/}
}
\usage{
dataRelated_activity
}
\description{
DATASET_DESCRIPTION
}
\examples{
{
knitr::kable(head(dataRelated_activity, 10))
}
}
\keyword{datasets}
