% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataSector}
\alias{dataSector}
\title{sector}
\format{
A data frame  
\describe{
  \item{\code{iati_identifier}}{character A globally unique identifier for the activity.}
  \item{\code{sector_vocabulary}}{character COLUMN_DESCRIPTION}
  \item{\code{sector_vocabulary_uri}}{logical COLUMN_DESCRIPTION}
  \item{\code{sector_code}}{character COLUMN_DESCRIPTION}
  \item{\code{sector_pct}}{character COLUMN_DESCRIPTION}
  \item{\code{sector_desc}}{character COLUMN_DESCRIPTION} 
  \item{\code{sector_vocabulary_name}}{character COLUMN_DESCRIPTION}
  \item{\code{sector_vocabulary_description}}{character COLUMN_DESCRIPTION} 
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/}
}
\usage{
dataSector
}
\description{
DATASET_DESCRIPTION
}
\examples{
{
knitr::kable(head(dataSector, 10))
}
}
\keyword{datasets}
