% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataTransaction}
\alias{dataTransaction}
\title{transaction}
\format{
A data frame  
\describe{
  \item{\code{iati_identifier}}{character COLUMN_DESCRIPTION}
  \item{\code{transaction_ref}}{character COLUMN_DESCRIPTION}
  \item{\code{transaction_humanitarian}}{logical COLUMN_DESCRIPTION}
  \item{\code{transaction_type_code}}{double COLUMN_DESCRIPTION}
  \item{\code{transaction_date}}{double COLUMN_DESCRIPTION}
  \item{\code{transaction_value_currency}}{character COLUMN_DESCRIPTION}
  \item{\code{transaction_value_date}}{double COLUMN_DESCRIPTION}
  \item{\code{transaction_value}}{double COLUMN_DESCRIPTION}
  \item{\code{transaction_description}}{character COLUMN_DESCRIPTION}
  \item{\code{transaction_provider_org_type}}{double COLUMN_DESCRIPTION}
  \item{\code{transaction_provider_org_ref}}{character COLUMN_DESCRIPTION}
  \item{\code{transaction_provider_org}}{character COLUMN_DESCRIPTION}
  \item{\code{transaction_aid_type_code_1}}{character COLUMN_DESCRIPTION}
  \item{\code{transaction_aid_type_vocabulary_1}}{double COLUMN_DESCRIPTION}
  \item{\code{transaction_aid_type_code_2}}{character COLUMN_DESCRIPTION}
  \item{\code{transaction_aid_type_vocabulary_2}}{double COLUMN_DESCRIPTION}
  \item{\code{transaction_value_USD}}{double COLUMN_DESCRIPTION}
  \item{\code{transaction_type_name}}{character COLUMN_DESCRIPTION}
  \item{\code{transaction_type_description}}{character COLUMN_DESCRIPTION}
  \item{\code{provider_org_type_name}}{character COLUMN_DESCRIPTION}
  \item{\code{provider_org_type_description}}{character COLUMN_DESCRIPTION}
  \item{\code{aid_type1_name}}{character COLUMN_DESCRIPTION}
  \item{\code{aid_type1_description}}{character COLUMN_DESCRIPTION}
  \item{\code{earmarking_name}}{character COLUMN_DESCRIPTION}
  \item{\code{earmarking_description}}{character COLUMN_DESCRIPTION}
  \item{\code{aid_type_vocabulary1_name}}{character COLUMN_DESCRIPTION}
  \item{\code{aid_type_vocabulary1_description}}{character COLUMN_DESCRIPTION}
  \item{\code{aid_type_vocabulary2_name}}{character COLUMN_DESCRIPTION}
  \item{\code{aid_type_vocabulary2_description}}{character COLUMN_DESCRIPTION} 
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/activity-standard/iati-activities/iati-activity/transaction/}
}
\usage{
dataTransaction
}
\description{
Transactions recording committed or actual funds flowing in or out of an aid activity.
}
\examples{
{
knitr::kable(head(dataTransaction, 10))
}
}
\keyword{datasets}
