% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_aid_type}
\alias{default_aid_type}
\title{DATASET_TITLE}
\format{
A data frame with 825 rows and 3 variables:
\describe{
  \item{\code{iati_identifier}}{character A globally unique identifier for the activity.}
  \item{\code{default_aid_type_code}}{character COLUMN_DESCRIPTION}
  \item{\code{default_aid_type_vocabulary}}{logical COLUMN_DESCRIPTION} 
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/}
}
\usage{
default_aid_type
}
\description{
DATASET_DESCRIPTION
}
\examples{
{
head(default_aid_type,10)
}
}
\keyword{datasets}
