% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{document_link}
\alias{document_link}
\title{DATASET_TITLE}
\format{
A data frame with 4722 rows and 7 variables:
\describe{
  \item{\code{iati_identifier}}{character A globally unique identifier for the activity.}
  \item{\code{document_url}}{character COLUMN_DESCRIPTION}
  \item{\code{document_format}}{character COLUMN_DESCRIPTION}
  \item{\code{document_title}}{character COLUMN_DESCRIPTION}
  \item{\code{document_category_code}}{character COLUMN_DESCRIPTION}
  \item{\code{document_language_code}}{character COLUMN_DESCRIPTION}
  \item{\code{document_date}}{character COLUMN_DESCRIPTION} 
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/}
}
\usage{
document_link
}
\description{
DATASET_DESCRIPTION
}
\examples{
{
head(document_link,10)
}
}
\keyword{datasets}
