% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_structure.R
\name{extract_structure}
\alias{extract_structure}
\title{Extract Plot Structure and Metadata}
\usage{
extract_structure(p)
}
\arguments{
\item{p}{A `ggplot` object.}
}
\value{
A list containing:
  \item{labels}{Title, subtitle, caption, and axis labels.}
  \item{ranges}{Exact x and y ranges for each panel (trained).}
  \item{guides}{Mapping of visuals (color/shape) to data values.}
  \item{geoms}{List of geometric layers used.}
}
\description{
This function  extracts visual metadata from a ggplot object 
by forcing a render build.
Unlike simple label extraction, this captures the "trained" ranges and
legend mappings that are actually displayed to the user.
}
\examples{
 library(ggplot2)
p <- ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point() +
  unhcrthemes::theme_unhcr(grid = "Y", axis = "X", axis_title = FALSE) +
  labs(
    title = "Vehicle Efficiency",
    subtitle = "Fuel consumption vs weight",
    caption = "Source: mtcars dataset"
  )

extract_structure(p)
}
