% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iati-package.R
\docType{package}
\name{iati}
\alias{iati-package}
\alias{iati}
\title{An Interface to the International Aid Transparency Initiative API}
\description{
The International Aid Transparency Initiatives (IATI) is a global aid
transparency standard and it makes information about aid spending easier to
access, use and understand using a unified open standard. The IATI
Application Programming Interface (API) enables IATI standard compliant
datasets to be easily retrieved. This package provides an interface to the
IATI API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/katilingban/iati}
  \item \url{https://katilingban.io/iati/}
  \item Report bugs at \url{https://github.com/katilingban/iati/issues}
}

}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID})

}
\keyword{internal}
