% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-list_registry_datasets.R
\name{iati_list_datasets}
\alias{iati_list_datasets}
\title{List all IATI Registry datasets}
\usage{
iati_list_datasets(
  base_url = "https://iatiregistry.org/api/action",
  limit = NULL,
  offset = NULL
)
}
\arguments{
\item{base_url}{The IATI API base URL. This is currently set to
https://iatiregistry.org/api/action.}

\item{limit}{The number of datasets to list per page. Only one page will be
returned at a time.}

\item{offset}{The offset to start returning packages from.}
}
\value{
A data.frame listing the IATI Registry datasets
}
\description{
List all IATI Registry datasets
}
\examples{
iati_list_datasets()

}
