% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iati_reshape.R
\name{iati_reshape}
\alias{iati_reshape}
\title{iati_reshape}
\usage{
iati_reshape(iati_file_url, extractors = NULL, verbose = TRUE)
}
\arguments{
\item{xml_iati}{xml_iati}

\item{folder_name}{folder_name within project}
}
\description{
Take anxml file and creates all the decomposed xls file out o it..
}
\examples{
#iati_reshape(
# iati_file_url = 'https://files.unhcr.org/en/reporting/unhcr-activities-2022.xml' , 
# folder_name = "data-raw")

#iati_reshape( iati_file_url = 'https://files.unhcr.org/en/reporting/unhcr-activities-2022.xml' ,   folder_name = "data-raw")
#iati_reshape( iati_file_url = 'https://files.unhcr.org/en/reporting/unhcr-activities-2023.xml' ,   folder_name = "data-raw")
#iati_reshape( iati_file_url = 'https://files.unhcr.org/en/reporting/unhcr-activities-2024.xml' ,   folder_name = "data-raw")
#iati_reshape( iati_file_url = 'https://files.unhcr.org/en/reporting/unhcr-activities-2025.xml' ,   folder_name = "data-raw")
#iati_reshape( iati_file_url = 'https://files.unhcr.org/en/reporting/unhcr-activities-2026.xml' ,   folder_name = "data-raw")
}
