% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{location}
\alias{location}
\title{DATASET_TITLE}
\format{
A data frame with 3664 rows and 15 variables:
\describe{
  \item{\code{iati_identifier}}{character COLUMN_DESCRIPTION}
  \item{\code{location_ref}}{character COLUMN_DESCRIPTION}
  \item{\code{location_reach}}{character COLUMN_DESCRIPTION}
  \item{\code{location_id_vocabulary}}{logical COLUMN_DESCRIPTION}
  \item{\code{location_id_code}}{logical COLUMN_DESCRIPTION}
  \item{\code{location_name}}{character COLUMN_DESCRIPTION}
  \item{\code{location_activity_descr}}{logical COLUMN_DESCRIPTION}
  \item{\code{location_adm_vocabulary}}{logical COLUMN_DESCRIPTION}
  \item{\code{location_adm_level}}{logical COLUMN_DESCRIPTION}
  \item{\code{location_adm_code}}{logical COLUMN_DESCRIPTION}
  \item{\code{location_lat}}{character COLUMN_DESCRIPTION}
  \item{\code{location_long}}{character COLUMN_DESCRIPTION}
  \item{\code{location_exactness_code}}{character COLUMN_DESCRIPTION}
  \item{\code{location_class_code}}{character COLUMN_DESCRIPTION}
  \item{\code{location_feature_designation_code}}{logical COLUMN_DESCRIPTION} 
}
}
\source{
\url{https://reporting.unhcr.org/iati}
}
\usage{
location
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
