% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapping_indicator}
\alias{mapping_indicator}
\title{mapping_indicator}
\format{
A data frame with 47 rows and 27 variables:
\describe{
  \item{\code{Indicator}}{character COLUMN_DESCRIPTION}
  \item{\code{Indicator_lab2}}{character COLUMN_DESCRIPTION}
  \item{\code{Area of work}}{character COLUMN_DESCRIPTION}
  \item{\code{theme}}{character COLUMN_DESCRIPTION}
  \item{\code{subtheme}}{character COLUMN_DESCRIPTION}
  \item{\code{Results_Level}}{character COLUMN_DESCRIPTION}
  \item{\code{area_id}}{character COLUMN_DESCRIPTION}
  \item{\code{Area}}{character COLUMN_DESCRIPTION}
  \item{\code{Area_id}}{character COLUMN_DESCRIPTION}
  \item{\code{Ind_id}}{character COLUMN_DESCRIPTION}
  \item{\code{Ind_seq}}{double COLUMN_DESCRIPTION}
  \item{\code{Show_As}}{character COLUMN_DESCRIPTION}
  \item{\code{Indicator_Code}}{character COLUMN_DESCRIPTION}
  \item{\code{Reverse}}{logical COLUMN_DESCRIPTION}
  \item{\code{threshold_red}}{double COLUMN_DESCRIPTION}
  \item{\code{threshold_orange}}{double COLUMN_DESCRIPTION}
  \item{\code{threshold_green}}{double COLUMN_DESCRIPTION}
  \item{\code{standard_direction}}{character COLUMN_DESCRIPTION}
  \item{\code{survey}}{character COLUMN_DESCRIPTION}
  \item{\code{RAS}}{double COLUMN_DESCRIPTION}
  \item{\code{STA}}{double COLUMN_DESCRIPTION}
  \item{\code{IDP}}{double COLUMN_DESCRIPTION}
  \item{\code{RET}}{double COLUMN_DESCRIPTION}
  \item{\code{OOC}}{double COLUMN_DESCRIPTION}
  \item{\code{all}}{double COLUMN_DESCRIPTION}
  \item{\code{DEN}}{character COLUMN_DESCRIPTION}
  \item{\code{max}}{character COLUMN_DESCRIPTION} 
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
mapping_indicator
}
\description{
Mapping Interpretation threshold for indicator values
}
\keyword{datasets}
