% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapping_result}
\alias{mapping_result}
\title{mapping_result}
\format{
A data frame with 107 rows and 2 variables:
\describe{
  \item{\code{result_title}}{character result_title}
  \item{\code{sector_rbm}}{character sector_rbm} 
}
}
\source{
\url{https://www.unhcr.org/what-we-do/build-better-futures/compass/}
}
\usage{
mapping_result
}
\description{
Mapping Results to UNHCR 2022 RBM Framework
}
\keyword{datasets}
