% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapping_sdg}
\alias{mapping_sdg}
\title{mapping_sdg}
\format{
A data frame with 26 rows and 2 variables:
\describe{
  \item{\code{area}}{character COLUMN_DESCRIPTION}
  \item{\code{sdg}}{character COLUMN_DESCRIPTION} 
}
}
\source{
\url{https://www.unhcr.org/what-we-do/build-better-futures/compass}#' @examples
{
knitr::kable(head(mapping_sdg, 26))
}
}
\usage{
mapping_sdg
}
\description{
Mapping UNHCR areas to UN SDG
}
\keyword{datasets}
