% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapping_sector}
\alias{mapping_sector}
\title{mapping_sector}
\format{
A data frame with 83 rows and 3 variables:
\describe{
  \item{\code{sector_desc}}{character COLUMN_DESCRIPTION} 
  \item{\code{sector_vocabulary_name}}{character COLUMN_DESCRIPTION}
  \item{\code{sector_rbm}}{character COLUMN_DESCRIPTION} 
  \item{\code{area_id}}{character COLUMN_DESCRIPTION} 
}
}
\source{
\url{https://www.unhcr.org/what-we-do/build-better-futures/compass}
}
\usage{
mapping_sector
}
\description{
Mapping Sector to UNHCR 2022 RBM Framework
}
\keyword{datasets}
