% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{participating_org}
\alias{participating_org}
\title{DATASET_TITLE}
\format{
A data frame with 13990 rows and 7 variables:
\describe{
  \item{\code{iati_identifier}}{character COLUMN_DESCRIPTION}
  \item{\code{participating_org_eng}}{character COLUMN_DESCRIPTION}
  \item{\code{participating_org_ref}}{character COLUMN_DESCRIPTION}
  \item{\code{participating_org_type}}{character COLUMN_DESCRIPTION}
  \item{\code{participating_org_role}}{character COLUMN_DESCRIPTION}
  \item{\code{participating_org_activity_id}}{logical COLUMN_DESCRIPTION}
  \item{\code{participating_org_crs_channel_code}}{logical COLUMN_DESCRIPTION} 
}
}
\source{
\url{https://reporting.unhcr.org/iati}
}
\usage{
participating_org
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
