% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_donor_concentration.R
\name{plot_donor_concentration}
\alias{plot_donor_concentration}
\title{Plot donor funding concentration}
\usage{
plot_donor_concentration(
  year = NULL,
  programme_lab = NULL,
  iati_identifier_ops = NULL,
  ctr_name = NULL,
  top_n = 10
)
}
\arguments{
\item{year}{A numeric value or a vector of numeric values to filter on year.}

\item{programme_lab}{A character vector corresponding to the name of the programme.}

\item{iati_identifier_ops}{A character vector corresponding to the name of the operation.}

\item{ctr_name}{A character vector corresponding to the name of the country.}

\item{top_n}{The number of top donors to show.}
}
\description{
This function creates a bar chart to visualize donor funding
concentration by showing the percentage of total funding from the top N donors.
}
