% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_donor_earmarking.R
\name{plot_donor_earmarking}
\alias{plot_donor_earmarking}
\title{Plot donor earmarking}
\usage{
plot_donor_earmarking(
  donor_name,
  year = NULL,
  by = "global",
  programme_lab = NULL,
  iati_identifier_ops = NULL,
  ctr_name = NULL
)
}
\arguments{
\item{donor_name}{The name of the donor to plot.}

\item{year}{A numeric value or a vector of numeric values to filter on year.}

\item{by}{The category to group by. One of "date", "global", "region", "country", "sector".}

\item{programme_lab}{A character vector corresponding to the name of the programme.}

\item{iati_identifier_ops}{A character vector corresponding to the name of the operation.}

\item{ctr_name}{A character vector corresponding to the name of the country.}
}
\description{
This function plots the earmarking of funding from a specific donor.
}
