% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_donor_funding_over_time.R
\name{plot_donor_funding_over_time}
\alias{plot_donor_funding_over_time}
\title{Plot donor funding over time}
\usage{
plot_donor_funding_over_time(
  donor_name,
  year = NULL,
  by = "global",
  programme_lab = NULL,
  iati_identifier_ops = NULL,
  ctr_name = NULL,
  top_n_countries = 10
)
}
\arguments{
\item{donor_name}{The name of the donor to plot.}

\item{year}{A numeric value or a vector of numeric values to filter on year.}

\item{by}{The category to group by. One of "global", "region", "country", "sector".}

\item{programme_lab}{A character vector corresponding to the name of the programme.}

\item{iati_identifier_ops}{A character vector corresponding to the name of the operation.}

\item{ctr_name}{A character vector corresponding to the name of the country.}

\item{top_n_countries}{The number of top countries to show when `by = "country"`.}
}
\description{
This function plots the funding from a specific donor over time,
with options to group the data by different categories.
}
