% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_donor_funding_volatility.R
\name{plot_donor_funding_volatility}
\alias{plot_donor_funding_volatility}
\title{Plot donor funding volatility}
\usage{
plot_donor_funding_volatility(
  donor_name,
  programme_lab = NULL,
  iati_identifier_ops = NULL,
  ctr_name = NULL
)
}
\arguments{
\item{donor_name}{The name of the donor to plot.}

\item{programme_lab}{A character vector corresponding to the name of the programme.}

\item{iati_identifier_ops}{A character vector corresponding to the name of the operation.}

\item{ctr_name}{A character vector corresponding to the name of the country.}
}
\description{
This function creates a bar chart showing the year-on-year
percentage change in funding for a specific donor.
}
