% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_donor_interest_profile.R
\name{plot_donor_interest_profile}
\alias{plot_donor_interest_profile}
\title{Plot donor interest profile}
\usage{
plot_donor_interest_profile(
  donor_name,
  by = "sector",
  year = NULL,
  programme_lab = NULL,
  iati_identifier_ops = NULL,
  ctr_name = NULL,
  top_n = 10
)
}
\arguments{
\item{donor_name}{The name of the donor to plot.}

\item{by}{The category to group by. One of "sector" or "region".}

\item{year}{A numeric value or a vector of numeric values to filter on year.}

\item{programme_lab}{A character vector corresponding to the name of the programme.}

\item{iati_identifier_ops}{A character vector corresponding to the name of the operation.}

\item{ctr_name}{A character vector corresponding to the name of the country.}

\item{top_n}{The number of top categories to show.}
}
\description{
This function creates a bar chart showing the breakdown of a
donor's funding by sector or region.
}
