% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_implementing_partner_earmarking.R
\name{plot_implementing_partner_earmarking}
\alias{plot_implementing_partner_earmarking}
\title{Plot implementing partner earmarking}
\usage{
plot_implementing_partner_earmarking(
  partner_name,
  year = NULL,
  by = "global",
  programme_lab = NULL,
  iati_identifier_ops = NULL,
  ctr_name = NULL
)
}
\arguments{
\item{partner_name}{The name of the implementing partner to plot.}

\item{year}{A numeric value or a vector of numeric values to filter on year.}

\item{by}{The category to group by. One of "date", "global", "region", "country", "sector".}

\item{programme_lab}{A character vector corresponding to the name of the programme.}

\item{iati_identifier_ops}{A character vector corresponding to the name of the operation.}

\item{ctr_name}{A character vector corresponding to the name of the country.}
}
\description{
This function plots the earmarking of funding for a specific implementing partner.
}
