% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{result}
\alias{result}
\title{DATASET_TITLE}
\format{
A data frame with 26667 rows and 24 variables:
\describe{
  \item{\code{iati_identifier}}{character COLUMN_DESCRIPTION}
  \item{\code{result_type}}{character COLUMN_DESCRIPTION}
  \item{\code{result_aggregation_status}}{character COLUMN_DESCRIPTION}
  \item{\code{result_title}}{character COLUMN_DESCRIPTION}
  \item{\code{result_desc}}{logical COLUMN_DESCRIPTION}
  \item{\code{result_indicator_measure}}{character COLUMN_DESCRIPTION}
  \item{\code{result_indicator_title}}{character COLUMN_DESCRIPTION}
  \item{\code{result_indicator_desc}}{character COLUMN_DESCRIPTION}
  \item{\code{result_indicator_baseline_value}}{character COLUMN_DESCRIPTION}
  \item{\code{result_indicator_baseline_year}}{character COLUMN_DESCRIPTION}
  \item{\code{result_indicator_baseline_date}}{logical COLUMN_DESCRIPTION}
  \item{\code{result_indicator_baseline_location_ref}}{character COLUMN_DESCRIPTION}
  \item{\code{result_indicator_baseline_dimension_1}}{character COLUMN_DESCRIPTION}
  \item{\code{result_indicator_baseline_dimension_value_1}}{character COLUMN_DESCRIPTION}
  \item{\code{result_indicator_baseline_dimension_2}}{character COLUMN_DESCRIPTION}
  \item{\code{result_indicator_baseline_dimension_value_2}}{character COLUMN_DESCRIPTION}
  \item{\code{result_indicator_period_start}}{character COLUMN_DESCRIPTION}
  \item{\code{result_indicator_period_end}}{character COLUMN_DESCRIPTION}
  \item{\code{result_indicator_target_value}}{character COLUMN_DESCRIPTION}
  \item{\code{result_indicator_target_location_ref}}{character COLUMN_DESCRIPTION}
  \item{\code{result_indicator_target_dimension_1}}{character COLUMN_DESCRIPTION}
  \item{\code{result_indicator_target_value_1}}{character COLUMN_DESCRIPTION}
  \item{\code{result_indicator_target_dimension_2}}{character COLUMN_DESCRIPTION}
  \item{\code{result_indicator_target_value_2}}{character COLUMN_DESCRIPTION} 
}
}
\source{
\url{https://reporting.unhcr.org/iati}
}
\usage{
result
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
