% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sector}
\alias{sector}
\title{DATASET_TITLE}
\format{
A data frame with 23793 rows and 6 variables:
\describe{
  \item{\code{iati_identifier}}{character COLUMN_DESCRIPTION}
  \item{\code{sector_vocabulary}}{character COLUMN_DESCRIPTION}
  \item{\code{sector_vocabulary_uri}}{logical COLUMN_DESCRIPTION}
  \item{\code{sector_code}}{character COLUMN_DESCRIPTION}
  \item{\code{sector_pct}}{character COLUMN_DESCRIPTION}
  \item{\code{sector_desc}}{character COLUMN_DESCRIPTION} 
}
}
\source{
\url{https://reporting.unhcr.org/iati}
}
\usage{
sector
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
