% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_budget_gap.R
\name{show_budget_gap}
\alias{show_budget_gap}
\title{show_budget_gap}
\usage{
show_budget_gap(
  year,
  programme_lab = NULL,
  iati_identifier_ops = NULL,
  ctr_name = NULL,
  weight_by = NULL
)
}
\arguments{
\item{year}{A numeric value corresponding to the first year of focus until the most recent year within the dataset.}

\item{programme_lab}{A character vector corresponding to the name of the programme.}

\item{iati_identifier_ops}{A character vector corresponding to the name of the operation.}

\item{ctr_name}{A character vector corresponding to the name of the country.}

\item{weight_by}{list of population group to weight the budget - 
"refugees", "asylum_seekers",
 "returned_refugees" "idps",
 "returned_idps", "stateless", 
 "ooc", "oip"  
default is null.}
}
\value{
a graph
}
\description{
UNHCR budgets are needs-based: it represents the total amount
of money that would be required were UNHCR to meet all of the needs that it is seeking to address.
}
\examples{
show_budget_gap(year = 2018, 
             ctr_name = "Brazil")
show_budget_gap(year = 2018, 
             ctr_name = "Brazil",
             weight_by = c("refugees", "oip"))
}
