% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_contributions.R
\name{show_contributions}
\alias{show_contributions}
\title{Title}
\usage{
show_contributions(
  year,
  programme_lab = NULL,
  iati_identifier_ops = NULL,
  ctr_name = NULL
)
}
\arguments{
\item{year}{A numeric value corresponding to the first year of focus until the most recent year within the dataset.}

\item{programme_lab}{A character vector corresponding to the name of the programme.}

\item{iati_identifier_ops}{A character vector corresponding to the name of the operation.}

\item{ctr_name}{A character vector corresponding to the name of the country.}
}
\value{
a graph
}
\description{
What’s the breakdown by contribution Type (Un-earmarked, Tightly earmarked, etc.) from Donor Funds by Year?
}
\examples{
knitr::kable(iati::dataTransaction |>
                dplyr::select(aid_type1_name, aid_type1_description)  |>
                dplyr::distinct() |>
                dplyr::filter(!(is.na(aid_type1_name))))
show_contributions(year = c(2018, 2019,2020, 2021, 2022, 2023), 
                   ctr_name = "Brazil") 
show_contributions(year = c(2018, 2019,2020, 2021, 2022, 2023), 
                   programme_lab = "The Americas") 
}
