% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_donor_earmarking_flexibility_over_time.R
\name{show_donor_earmarking_flexibility_over_time}
\alias{show_donor_earmarking_flexibility_over_time}
\title{Plot donor earmarking flexibility over time}
\usage{
show_donor_earmarking_flexibility_over_time(
  donor_name,
  programme_lab = NULL,
  iati_identifier_ops = NULL,
  ctr_name = NULL
)
}
\arguments{
\item{donor_name}{The name of the donor to plot.}

\item{programme_lab}{A character vector corresponding to the name of the programme.}

\item{iati_identifier_ops}{A character vector corresponding to the name of the operation.}

\item{ctr_name}{A character vector corresponding to the name of the country.}
}
\description{
This function creates a stacked bar chart showing the proportion
of each earmarking category for a specific donor over several years.
}
\examples{
# Earmarking Flexibility Over Time
show_donor_earmarking_flexibility_over_time(donor_name = "Private donors",
                                            ctr_name = "Syria")


}
