% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_donor_earmarking_profile.R
\name{show_donor_earmarking_profile}
\alias{show_donor_earmarking_profile}
\title{Show donor earmarking profiles}
\usage{
show_donor_earmarking_profile(
  year,
  ctr_name = NULL,
  programme_lab = NULL,
  iati_identifier_ops = NULL,
  top_n = 10
)
}
\arguments{
\item{year}{A numeric value for the year.}

\item{ctr_name}{A character vector for the country name.}

\item{programme_lab}{A character vector for the programme name.}

\item{iati_identifier_ops}{A character vector for the operation ID.}

\item{top_n}{The number of top donors to display.}
}
\description{
This function creates a stacked bar chart to show the earmarking
profile of the top N donors.
}
\examples{
show_donor_earmarking_profile(year = 2025, ctr_name = "Brazil", top_n = 10)
}
