% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_donors.R
\name{show_donors}
\alias{show_donors}
\title{Ploting Donors}
\usage{
show_donors(
  year,
  iati_identifier_ctr_region,
  TransactionType = "Incoming Commitment"
)
}
\arguments{
\item{year}{year to select starting from 2016 - could be one year or a list}

\item{iati_identifier_ctr_region}{UNHCR Region}

\item{TransactionType}{Transaction type}
}
\value{
a graph
}
\description{
Ploting Donors
}
\examples{

knitr::kable( codeTransactionType |> dplyr::select(name,  description) )
show_donors(year = 2018,
           iati_identifier_ctr_region = "The Americas",
            TransactionType = "Incoming Commitment" )
}
