% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_goal_sdg.R
\name{show_goal_sdg}
\alias{show_goal_sdg}
\title{show_goal_sdg}
\usage{
show_goal_sdg(
  year,
  programme_lab = NULL,
  iati_identifier_ops = NULL,
  ctr_name = NULL
)
}
\arguments{
\item{year}{A numeric value or a list of value.}

\item{programme_lab}{A character vector corresponding to the name of the programme.}

\item{iati_identifier_ops}{A character vector corresponding to the name of the operation.}

\item{ctr_name}{A character vector corresponding to the name of the country.}
}
\value{
a graph
}
\description{
Show resource allocation in line with SDG goal
}
\examples{
show_goal_sdg(year =  c(  2020, 2021, 2022), 
             ctr_name = "Brazil")
}
