% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_implementers.R
\name{show_implementers}
\alias{show_implementers}
\title{show_implementers Title}
\usage{
show_implementers(year, geo, sector, metric)
}
\arguments{
\item{year}{year to select starting from 2016 - could be one year or a list}

\item{geo}{Region or Country or Site}

\item{sector}{Sectors}

\item{metric}{Metrics could be USD or Number of project}
}
\value{
1
}
\description{
show_implementers Title
}
\examples{
show_implementers()
}
