% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_indicators.R
\name{show_indicators}
\alias{show_indicators}
\title{show_indicators}
\usage{
show_indicators(
  year,
  programme_lab = NULL,
  iati_identifier_ops = NULL,
  ctr_name = NULL,
  result_type_name = "Outcome",
  type = "deviation"
)
}
\arguments{
\item{year}{A numeric value corresponding to the first year of focus until the most recent year within the dataset.}

\item{programme_lab}{A character vector corresponding to the name of the programme.}

\item{iati_identifier_ops}{A character vector corresponding to the name of the operation.}

\item{ctr_name}{A character vector corresponding to the name of the country.}

\item{result_type_name}{either  "Impact"  "Outcome" "Output" - default is  "Outcome"}

\item{type}{either "deviation" ,  "progress"  or "gap"
"deviation" showing difference between target and actual - or 
"progress"  showing difference between baseline and actual  - or
"gap" showing difference between green acceptable threshold and actual. 
gap analysis is only displayed for outcome indicators published after 2022}
}
\value{
a graph
}
\description{
How much indicators evolve over time against thresholds?
}
\examples{
show_indicators(year = 2022,  
             ctr_name = "Brazil",
             result_type_name = "Outcome",
             type = "deviation"
             )
show_indicators(year = 2022,  
             ctr_name = "Brazil",
             result_type_name = "Outcome",
             type = "progress"
             )
show_indicators(year = 2022,  
             ctr_name = "Brazil",
             result_type_name = "Outcome",
             type = "gap"
             )
show_indicators(year = 2022,  
             ctr_name = "Brazil",
             result_type_name = "Impact",
             type = "deviation"
             )
show_indicators(year = 2022,  
             ctr_name = "Brazil",
             result_type_name = "Impact",
             type = "progress"
             )
show_indicators(year = 2019,  
             ctr_name = "Brazil",
             result_type_name = "Output",
             type = "deviation"
             )
show_indicators(year = 2019,  
             ctr_name = "Brazil",
              result_type_name = "Output",
             type = "progress"
             )
}
