% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_indicators_time.R
\name{show_indicators_time}
\alias{show_indicators_time}
\title{show_indicators_time}
\usage{
show_indicators_time(
  year,
  programme_lab = NULL,
  iati_identifier_ops = NULL,
  ctr_name = NULL,
  result_type_name = "Outcome",
  type = "deviation"
)
}
\arguments{
\item{year}{A numeric value corresponding to the first year of focus until the most recent year within the dataset.}

\item{programme_lab}{A character vector corresponding to the name of the programme.}

\item{iati_identifier_ops}{A character vector corresponding to the name of the operation.}

\item{ctr_name}{A character vector corresponding to the name of the country.}

\item{result_type_name}{either  "Impact"  "Outcome" "Output" - default is  "Outcome"}

\item{type}{"deviation" showing difference between target and actual - or 
"progress"  showing difference between baseline and actual}
}
\value{
a graph
}
\description{
How much indicators evolve over time against thresholds?
}
\examples{
show_indicators_time(year = 2020,  
             ctr_name = "Brazil",
              result_type_name = "Outcome",
             type = "deviation"
             )
show_indicators_time(year =  2020, 
             ctr_name = "Brazil",
              result_type_name = "Impact",
             type = "deviation"
             )
show_indicators_time(year =  2020,  
             ctr_name = "Brazil",
              result_type_name = "Output",
             type = "deviation"
             )
show_indicators_time(year =  2020, 
             ctr_name = "Brazil",
              result_type_name = "Outcome",
             type = "progress"
             )
show_indicators_time(year =  2020, 
             ctr_name = "Brazil",
              result_type_name = "Impact",
             type = "progress"
             )
show_indicators_time(year =  2020, 
             ctr_name = "Brazil",
              result_type_name = "Output",
             type = "progress"
             )
}
