% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_outcome_result.R
\name{show_outcome_result}
\alias{show_outcome_result}
\title{Title}
\usage{
show_outcome_result(
  year,
  programme_lab = NULL,
  iati_identifier_ops = NULL,
  ctr_name = NULL,
  outcome
)
}
\arguments{
\item{year}{A numeric value corresponding to the first year of focus until the most recent year within the dataset.}

\item{programme_lab}{A character vector corresponding to the name of the programme.}

\item{iati_identifier_ops}{A character vector corresponding to the name of the operation.}

\item{ctr_name}{A character vector corresponding to the name of the country.}

\item{outcome}{any of:
"OA1: Access to Territory, Reg. and Documentation", 
"OA2: Status Determination",
"OA3: Protection Policy and Law", 
"OA4: Sexual and Gender-based Violence",
"OA5: Child Protection",
"OA6: Safety and Access to Justice", 
"OA7: Community Engagement and Women's Empowerment", 
"OA8: Well-Being and Basic Needs", 
"OA9: Sustainable Housing and Settlements", 
"OA10: Healthy Lives" ,
"OA11: Education",  
"OA12: Clean Water, Sanitation and Hygiene",
"OA13: Self Reliance, Economic Inclusion and Livelihoods",
"OA14: Voluntary Repatriation and Sustainable Reintegration",
"OA15: Resettlement and Complementary Pathways", 
"OA16: Local Integration and other Local Solutions"}
}
\value{
a graph
}
\description{
Title
}
\examples{
show_outcome_result(year =  c(  2020, 2021, 2022), 
             ctr_name = "Brazil",
             outcome = "OA2: Status Determination")
show_outcome_result(year =  c(  2020, 2021, 2022), 
             ctr_name = "Brazil",
             outcome = "OA9: Sustainable Housing and Settlements")
show_outcome_result(year =  c(  2020, 2021, 2022), 
             ctr_name = "Brazil",
             outcome = "OA8: Well-Being and Basic Needs")
show_outcome_result(year =  c(  2020, 2021, 2022), 
             ctr_name = "Brazil",
             outcome = "OA4: Sexual and Gender-based Violence")
}
