% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_partnership.R
\name{show_partnership}
\alias{show_partnership}
\title{show_partnership}
\usage{
show_partnership(
  year,
  programme_lab = NULL,
  iati_identifier_ops = NULL,
  ctr_name = NULL
)
}
\arguments{
\item{year}{A numeric value corresponding to the year to be displayed}

\item{programme_lab}{A character vector corresponding to the name of the programme.}

\item{iati_identifier_ops}{A character vector corresponding to the name of the operation.}

\item{ctr_name}{A character vector corresponding to the name of the country.}
}
\value{
a graph
}
\description{
How organisations partner together?
}
\examples{
show_partnership(year = 2022,
           ctr_name = "Brazil" ) 
}
