% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_sectors_rbm.R
\name{show_sectors_rbm}
\alias{show_sectors_rbm}
\title{show_sectors_rbm}
\usage{
show_sectors_rbm(
  year,
  programme_lab = NULL,
  iati_identifier_ops = NULL,
  ctr_name = NULL
)
}
\arguments{
\item{year}{A numeric value or a list of value.}

\item{programme_lab}{A character vector corresponding to the name of the programme.}

\item{iati_identifier_ops}{A character vector corresponding to the name of the operation.}

\item{ctr_name}{A character vector corresponding to the name of the country.}
}
\value{
a graph
}
\description{
What are the most funded sectors per country based on new RBM framework
}
\examples{
show_sectors_rbm( year =  c(2017, 2018, 2019, 2020, 2021, 2022), 
             ctr_name = "Brazil")
}
