% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_top_donors.R
\name{show_top_donors}
\alias{show_top_donors}
\title{show_top_donors}
\usage{
show_top_donors(
  year,
  programme_lab = NULL,
  iati_identifier_ops = NULL,
  ctr_name = NULL,
  top_n = 5
)
}
\arguments{
\item{year}{year to select starting from 2016 - could be one year or a list}

\item{programme_lab}{A character vector corresponding to the name of the programme.}

\item{iati_identifier_ops}{A character vector corresponding to the name of the operation.}

\item{ctr_name}{A character vector corresponding to the name of the country.}

\item{top_n}{top n donors to show - the rest being lumped together}
}
\value{
a graph
}
\description{
Who are the main donors by country in terms of transaction?
}
\examples{
show_top_donors(year = 2022,
           ctr_name = "Brazil", 
           top_n = 5)
show_top_donors(year = 2020,
           ctr_name = "Brazil", 
           top_n = 10)
}
