% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_transaction_flow.R
\name{show_transaction_flow}
\alias{show_transaction_flow}
\title{show_transaction_flow}
\usage{
show_transaction_flow(
  year,
  programme_lab = NULL,
  iati_identifier_ops = NULL,
  ctr_name = NULL,
  top_n = 10
)
}
\arguments{
\item{year}{A numeric value corresponding to the first year of focus until the most recent year within the dataset.}

\item{programme_lab}{A character vector corresponding to the name of the programme.}

\item{iati_identifier_ops}{A character vector corresponding to the name of the operation.}

\item{ctr_name}{A character vector corresponding to the name of the country.}

\item{top_n}{top n donors to show - the rest being lumped together default is 10}
}
\value{
a graph
}
\description{
Flow diagramme of transations
}
\examples{
show_transaction_flow(year = c(2020, 2021, 2022, 2023), 
                   ctr_name = "Brazil")

show_transaction_flow(year = 2023, 
                   ctr_name = "Brazil")
}
