% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_compare.R
\name{template_compare}
\alias{template_compare}
\title{Generate a summary powerpoint}
\usage{
template_compare(year = 2022, folder = "Compare")
}
\arguments{
\item{year}{A numeric value or a vector of numeric value to filter on year. Note that data pre-2022 are using a different set of indicators}

\item{folder}{folder within your project where to put the generated report. 
Folder will be created if it does not exist}

\item{ctr_name}{A character vector corresponding to the name of the country.}
}
\value{
nothing the file for the report is generated
}
\description{
Generate a summary powerpoint
}
\examples{
# iati::template_compare(year = 2022, folder = "dev/Result")
}
