% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_prez.R
\name{template_prez}
\alias{template_prez}
\title{Generate a summary powerpoint}
\usage{
template_prez(year = 2022, ctr_name, folder = "Prez")
}
\arguments{
\item{year}{A numeric value or a vector of numeric value to filter on year. Note that data pre-2022 are using a different set of indicators}

\item{ctr_name}{A character vector corresponding to the name of the country.}

\item{folder}{folder within your project where to put the generated report. 
Folder will be created if it does not exist}
}
\value{
nothing the file for the report is generated
}
\description{
Generate a summary powerpoint
}
\examples{

## generate for one country
# iati::template_prez(year = 2022, 
#                     ctr_name = "Brazil",
#                     folder = "dev/Prez")

# ## Generate for all operation specific region
# region <- "The Americas"
# 
# countries <- iati::dataActivity |> 
#        dplyr::filter( unhcr_region ==  region) |>
#        dplyr::select(ctr_name) |>
#        dplyr::distinct() |>
#        dplyr::pull()
# 
# for ( ctr in countries) {
#     cat(paste0(ctr, "\n"))
#     iati::template_prez(year = 2022, 
#                     ctr_name = ctr,
#                     folder = "dev/Prez")  }
}
