% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transaction}
\alias{transaction}
\title{DATASET_TITLE}
\format{
A data frame with 30752 rows and 17 variables:
\describe{
  \item{\code{iati_identifier}}{character COLUMN_DESCRIPTION}
  \item{\code{transaction_ref}}{character COLUMN_DESCRIPTION}
  \item{\code{transaction_humanitarian}}{logical COLUMN_DESCRIPTION}
  \item{\code{transaction_type_code}}{character COLUMN_DESCRIPTION}
  \item{\code{transaction_date}}{character COLUMN_DESCRIPTION}
  \item{\code{transaction_value_currency}}{character COLUMN_DESCRIPTION}
  \item{\code{transaction_value_date}}{character COLUMN_DESCRIPTION}
  \item{\code{transaction_value}}{character COLUMN_DESCRIPTION}
  \item{\code{transaction_description}}{character COLUMN_DESCRIPTION}
  \item{\code{transaction_provider_org_type}}{character COLUMN_DESCRIPTION}
  \item{\code{transaction_provider_org_ref}}{character COLUMN_DESCRIPTION}
  \item{\code{transaction_provider_org}}{character COLUMN_DESCRIPTION}
  \item{\code{transaction_aid_type_code_1}}{character COLUMN_DESCRIPTION}
  \item{\code{transaction_aid_type_vocabulary_1}}{character COLUMN_DESCRIPTION}
  \item{\code{transaction_aid_type_code_2}}{character COLUMN_DESCRIPTION}
  \item{\code{transaction_aid_type_vocabulary_2}}{character COLUMN_DESCRIPTION}
  \item{\code{transaction_value_USD}}{character COLUMN_DESCRIPTION} 
}
}
\source{
\url{https://reporting.unhcr.org/iati}
}
\usage{
transaction
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
