% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codeActivityStatus}
\alias{codeActivityStatus}
\title{codeActivityStatus}
\format{
A data frame with 6 rows and 6 variables:
\describe{
  \item{\code{code}}{double COLUMN_DESCRIPTION}
  \item{\code{name}}{character COLUMN_DESCRIPTION}
  \item{\code{description}}{character COLUMN_DESCRIPTION}
  \item{\code{category}}{logical COLUMN_DESCRIPTION}
  \item{\code{url}}{logical COLUMN_DESCRIPTION}
  \item{\code{status}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/codelists/}
}
\usage{
codeActivityStatus
}
\description{
Lifecycle status of the activity from pipeline to completion
}
\examples{
{
head(codeActivityStatus,10)
}
}
\keyword{datasets}
