% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataBudget}
\alias{dataBudget}
\title{dataBudget}
\format{

}
\source{
\url{https://iatistandard.org/en/iati-standard/203/}
}
\usage{
dataBudget
}
\description{
Display budget by operation
}
\examples{
{
knitr::kable(head(dataBudget, 10))
}
}
\keyword{datasets}
