% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataBudget}
\alias{dataBudget}
\title{dataBudget}
\format{
A data frame with 840 rows and 8 variables:
\describe{
  \item{\code{iati_identifier}}{character A globally unique identifier for the activity.}
  \item{\code{budget_type}}{character COLUMN_DESCRIPTION}
  \item{\code{budget_status}}{character COLUMN_DESCRIPTION}
  \item{\code{budget_period_start}}{character COLUMN_DESCRIPTION}
  \item{\code{budget_period_end}}{character COLUMN_DESCRIPTION}
  \item{\code{budget_currency}}{character COLUMN_DESCRIPTION}
  \item{\code{budget_value_date}}{character COLUMN_DESCRIPTION}
  \item{\code{budget_value}}{character COLUMN_DESCRIPTION} 
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/}
}
\usage{
dataBudget
}
\description{
DATASET_DESCRIPTION
}
\examples{
{
knitr::kable(head(dataBudget, 10))
}
}
\keyword{datasets}
