% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataBudget}
\alias{dataBudget}
\title{dataBudget}
\format{
A data frame with 840 rows and 8 variables:
\describe{
  \item{\code{iati_identifier}}{character A globally unique identifier for the activity.}
  \item{\code{budget_type}}{character budget_type}
  \item{\code{budget_status}}{character budget_status}
  \item{\code{budget_period_start}}{character budget_period_start}
  \item{\code{budget_period_end}}{character budget_period_end}
  \item{\code{budget_currency}}{character budget_currency}
  \item{\code{budget_value_date}}{character budget_value_date}
  \item{\code{budget_value}}{character budget_value} 
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/}
}
\usage{
dataBudget
}
\description{
Display budget by operation
}
\examples{
{
knitr::kable(head(dataBudget, 10))
}
}
\keyword{datasets}
