% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataDefault_aid_type}
\alias{dataDefault_aid_type}
\title{dataDefault_aid_type}
\format{
A data frame  
\describe{
  \item{\code{iati_identifier}}{character A globally unique identifier for the activity.}
  \item{\code{default_aid_type_code}}{character aid_type_code}
  \item{\code{default_aid_type_vocabulary}}{logical aid_type_vocabulary} 
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/}
}
\usage{
dataDefault_aid_type
}
\description{
Default_aid_type
}
\examples{
{
knitr::kable(head(dataDefault_aid_type, 10))
}
}
\keyword{datasets}
