% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{humanitarian_scope}
\alias{humanitarian_scope}
\title{DATASET_TITLE}
\format{
A data frame with 1578 rows and 8 variables:
\describe{
  \item{\code{iati_identifier}}{character A globally unique identifier for the activity.}
  \item{\code{humanitarian_scope_type}}{character COLUMN_DESCRIPTION}
  \item{\code{humanitarian_scope_vocabulary}}{character COLUMN_DESCRIPTION}
  \item{\code{humanitarian_scope_vocabulary_uri}}{character COLUMN_DESCRIPTION}
  \item{\code{humanitarian_scope_code}}{character COLUMN_DESCRIPTION}
  \item{\code{humanitarian_scope_pct}}{logical COLUMN_DESCRIPTION}
  \item{\code{humanitarian_scope_desc_eng}}{character COLUMN_DESCRIPTION}
  \item{\code{humanitarian_scope_desc_fr}}{character COLUMN_DESCRIPTION} 
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/}
}
\usage{
humanitarian_scope
}
\description{
DATASET_DESCRIPTION
}
\examples{
{
knitr::kable(head(humanitarian_scope, 10))
}
}
\keyword{datasets}
