% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iati-package.R
\docType{package}
\name{iati-package}
\alias{iati}
\alias{iati-package}
\title{iati: UNHCR IATI Data Crunching}
\description{
This package includes UNHCR IATI data organised in tabular format together with a chat library.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://edouard-legoupil.github.io/iati/}
  \item Report bugs at \url{https://github.com/edouard-legoupil/iati/issues}
}

}
\author{
\strong{Maintainer}: Edouard Legoupil \email{legoupil@unhcr.org}

Authors:
\itemize{
  \item Matheus Soldi Hardt \email{matheus@unhcr.org}
}

Other contributors:
\itemize{
  \item UNHCR \email{iati@unhcr.org} [copyright holder]
}

}
\keyword{internal}
