% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iati-package.R
\docType{package}
\name{iati-package}
\alias{iati}
\alias{iati-package}
\title{iati: UNHCR IATI Data Crunching}
\description{
This is a data package with all frame organised and reshapped for easy consumptions in R Please see https://reporting.unhcr.org/iati
}
\seealso{
Useful links:
\itemize{
  \item \url{https://unhcr-americas.github.io/iati/docs/}
  \item Report bugs at \url{https://github.com/unhcr-americas/iati/issues}
}

}
\keyword{internal}
