% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{related_activity}
\alias{related_activity}
\title{DATASET_TITLE}
\format{
A data frame with 2950 rows and 3 variables:
\describe{
  \item{\code{iati_identifier}}{character A globally unique identifier for the activity.}
  \item{\code{related_activity_ref}}{character COLUMN_DESCRIPTION}
  \item{\code{related_activity_type}}{character COLUMN_DESCRIPTION} 
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/}
}
\usage{
related_activity
}
\description{
DATASET_DESCRIPTION
}
\examples{
{
knitr::kable(head(related_activity, 10))
}
}
\keyword{datasets}
