% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_donors.R
\name{show_donors}
\alias{show_donors}
\title{Ploting Donors}
\usage{
show_donors(x, na.rm = TRUE)
}
\arguments{
\item{year}{year to select starting from 2016 - could be one year or a list}

\item{geo}{Region or Country or Site}

\item{sector}{Sectors}

\item{metric}{Metrics could be USD or Number of project}
}
\value{
Median of vector x
}
\description{
Ploting Donors
}
\examples{
show_donors(1:12)
}
