% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_donors.R
\name{show_donors}
\alias{show_donors}
\title{Ploting Donors}
\usage{
show_donors(year, geo, transtype)
}
\arguments{
\item{year}{year to select starting from 2016 - could be one year or a list}

\item{geo}{Region or Country or Site}

\item{transtype}{Transaction type}
}
\description{
Ploting Donors
}
\examples{
show_donors(year = 2020,
            geo = "The Americas",
            transtype = "11" )
# ,  ## "3"  "4"  "11"           metric = "USD"
}
