% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_donors.R
\name{show_donors}
\alias{show_donors}
\title{show_donors}
\usage{
show_donors(
  year,
  programme_lab = NULL,
  iati_identifier_ops = NULL,
  ctr_name = NULL,
  transaction_type_name = "Incoming Commitment"
)
}
\arguments{
\item{year}{year to select starting from 2016 - could be one year or a list}

\item{programme_lab}{A character vector corresponding to the name of the programme.}

\item{iati_identifier_ops}{A character vector corresponding to the name of the operation.}

\item{ctr_name}{A character vector corresponding to the name of the country.}

\item{transaction_type_name}{Transaction type - default is "Incoming Commitment" ,
can also be   "Disbursement", or "Expenditure"}
}
\value{
a graph
}
\description{
Who are the main donors by country in terms of number of projects and/or total budget?
}
\examples{

knitr::kable( codeTransactionType |> dplyr::select(name,  description) )

show_donors(year = 2018,
           programme_lab = "The Americas",
           transaction_type_name = "Incoming Commitment" )

show_donors(year = 2018,
           ctr_name = "Brazil",
           transaction_type_name = "Incoming Commitment" )


show_donors(year = 2018,
           programme_lab = "The Americas",
           transaction_type_name = "Disbursement" )


show_donors(year = 2018,
           programme_lab = "The Americas",
           transaction_type_name = "Expenditure" )

}
