% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_donors.R
\name{show_donors}
\alias{show_donors}
\title{Ploting Donors}
\usage{
show_donors(
  thisyear,
  thisprogrammme_lab,
  thistransaction_type_name = "Incoming Commitment"
)
}
\arguments{
\item{thisyear}{year to select starting from 2016 - could be one year or a list}

\item{thisprogrammme_lab}{programme_lab}

\item{thistransaction_type_name}{Transaction type}
}
\value{
a graph
}
\description{
Ploting Donors
}
\examples{

knitr::kable( codeTransactionType |> dplyr::select(name,  description) )

show_donors(thisyear = 2018,
           thisprogrammme_lab = "The Americas",
           thistransaction_type_name = "Incoming Commitment" )


show_donors(thisyear = 2018,
           thisprogrammme_lab = "The Americas",
           thistransaction_type_name = "Disbursement" )

}
