% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_expenditure.R
\name{show_expenditure}
\alias{show_expenditure}
\title{show_expenditure}
\usage{
show_expenditure(
  year,
  programme_lab = NULL,
  iati_identifier_ops = NULL,
  ctr_name = NULL,
  weight_by = NULL
)
}
\arguments{
\item{year}{A numeric value corresponding to the first year of focus until the most recent year within the dataset.}

\item{programme_lab}{A character vector corresponding to the name of the programme.}

\item{iati_identifier_ops}{A character vector corresponding to the name of the operation.}

\item{ctr_name}{A character vector corresponding to the name of the country.}

\item{weight_by}{list of population group to weight the budget - "all" or any of
"refugees", "asylum_seekers",
 "returned_refugees" "idps",
 "returned_idps", "stateless", 
 "ooc", "oip"  
default is null.}
}
\value{
a graph
}
\description{
How much expenditures compare to the initial budget (potentially weighted by #)?
}
\examples{
show_expenditure(year = c(2018, 2019,2020, 2021, 2022, 2023), 
             ctr_name = "Brazil")

show_expenditure(year = c(2018, 2019,2020, 2021, 2022, 2023), 
             ctr_name = "Brazil",
             weight_by = c("refugees", "oip"))


show_expenditure(year = c(2018, 2019,2020, 2021, 2022, 2023),
             ctr_name = "Argentina",
             weight_by =  "all" )

}
