% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_implementers.R
\name{show_implementers}
\alias{show_implementers}
\title{Ploting implementers}
\usage{
show_implementers(year, geo, transtype)
}
\arguments{
\item{year}{year to select starting from 2016 - could be one year or a list}

\item{geo}{Region or Country or Site}

\item{transtype}{Transaction type}
}
\description{
Ploting implementers
}
\examples{
show_implementers(year = 2020,
            geo = "The Americas",
            transtype = "11" ) 
}
